requirements.txt
# Spotify API
spotipy==2.23.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.15.0

# Model Explainability
shap==0.44.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
jupyter==1.0.0
ipykernel==6.27.1

# Data validation
pydantic==2.5.3
